Several PXE fixes/additions for both EFI and BIOS - Implement EFI PXE boot! - Fix SMP trampoline for BIOS PXE boot - Add PXE EFI declarations - Align stack to 16 bytes before entering bootloader code - Drastically improve FAT seek performance - Make boot_pread return ssize_t instead of int - Fix broken module loader - Make filesystem device lists and partition lists unlimited - Stop using BIOS data area for VGA I/O base address - Properly construct and destruct fat32 file handles - Fix handling of BPB serial, it is 32 bits - Fix a couple of major bugs in vector<T> - Add zero_init_t base class that does zero initialized malloc/free for new/delete - Stop freeezing in cpu_debug_break

+++ boot/elf64decl.h
@@ -7,7 +7,7 @@ typedef uint64_t Elf64_Addr;
 typedef int64_t Elf64_Off;
 typedef uint16_t Elf64_Half;
 typedef uint32_t Elf64_Word;
-typedef int64_t Elf64_Sword;
+typedef int32_t Elf64_Sword;
 typedef uint64_t Elf64_Xword;
 typedef int64_t Elf64_Sxword;
